+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Enhancements
 
 - Setter/Remover method with `SharedPreferences.Editor#commit()` is now available. You need to enable it with `@Key(needCommitMethod = true)`
+- Add overload method for Getter that you can pass default value as parameter
 
 ## 1.2.0 - 2016/03/12